Simplification in family [gte_file], which now includes only [gte_file_open] and [gte_file_write].
Use of [g_signal_connect_data] instead of [g_signal_connect] to connect signal [changed] from text buffer to its handler. This signal is sent when a file is opened and the buffer is defined, even if the file was not modified by the user. [g_signal_connect_data] returns the handler id [gte_textview_signal_change_handler], which is used in [gte_file_open] to block signal [changed] before the buffer is redefined, so it does not change the title.